:broom: Add specs for container.rb, relocate files
