+++ app/transactions/bulkrax/container.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+require 'dry/container'
+
+module Bulkrax
+  class Container
+    extend Dry::Container::Mixin
+
+    namespace "work_resource" do |ops|
+      ops.register "create_with_bulk_behavior" do
+        steps = Hyrax::Transactions::WorkCreate::DEFAULT_STEPS.dup
+        steps[steps.index("work_resource.add_file_sets")] = "work_resource.add_bulkrax_files"
+
+        Hyrax::Transactions::WorkCreate.new(steps: steps)
+      end
+
+      ops.register "update_with_bulk_behavior" do
+        steps = Hyrax::Transactions::WorkUpdate::DEFAULT_STEPS.dup
+        steps[steps.index("work_resource.add_file_sets")] = "work_resource.add_bulkrax_files"
+
+        Hyrax::Transactions::WorkUpdate.new(steps: steps)
+      end
+
+      # TODO: uninitialized constant Bulkrax::Container::InlineUploadHandler
+      # ops.register "add_file_sets" do
+      #   Hyrax::Transactions::Steps::AddFileSets.new(handler: InlineUploadHandler)
+      # end
+
+      ops.register "add_bulkrax_files" do
+        Bulkrax::Steps::AddFiles.new
+      end
+    end
+  end
+end
+Hyrax::Transactions::Container.merge(Bulkrax::Container)
