+++ app/transactions/bulkrax/transactions.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-require 'bulkrax/transactions/container'
+require 'bulkrax/container'
 
 module Bulkrax
   ##
